Separate scoping functionality into its own module
